revert test_tweak separation.
